Remove hard hearing dependency
