+++ addons/compat_gm/compat_gm_hearing/CfgWeapons.hpp
@@ -1,0 +1,125 @@
+class CfgWeapons {
+    // HELMETS
+    class gm_ge_headgear_headset_crew_base;
+    class gm_ge_headgear_headset_crew_oli: gm_ge_headgear_headset_crew_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    class gm_ge_headgear_sph4_base;
+    class gm_ge_headgear_sph4_oli: gm_ge_headgear_sph4_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    class gm_pl_headgear_wz63_base;
+    class gm_pl_army_headgear_wz63_oli: gm_pl_headgear_wz63_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_pl_army_headgear_wz63_net_oli: gm_pl_headgear_wz63_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    class gm_ge_headgear_crewhat_80_base;
+    class gm_ge_headgear_crewhat_80_blk: gm_ge_headgear_crewhat_80_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    class gm_gc_headgear_crewhat_80_base;
+    class gm_gc_army_headgear_crewhat_80_blk: gm_gc_headgear_crewhat_80_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    class gm_gc_headgear_zsh3_base;
+    class gm_gc_headgear_zsh3_wht: gm_gc_headgear_zsh3_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_gc_headgear_zsh3_blu: gm_gc_headgear_zsh3_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_gc_headgear_zsh3_orn: gm_gc_headgear_zsh3_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    class gm_ge_headgear_beret_crew_base;
+    class gm_ge_headgear_beret_crew_blk: gm_ge_headgear_beret_crew_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    class gm_ge_headgear_beret_crew_bdx;
+    class gm_ge_headgear_beret_crew_grn;
+    class gm_ge_headgear_beret_crew_red;
+
+    class gm_ge_headgear_beret_crew_red_antiair: gm_ge_headgear_beret_crew_red {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_ge_headgear_beret_crew_red_artillery: gm_ge_headgear_beret_crew_red {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_ge_headgear_beret_crew_red_engineer: gm_ge_headgear_beret_crew_red {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_ge_headgear_beret_crew_red_maintenance: gm_ge_headgear_beret_crew_red {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_ge_headgear_beret_crew_grn_mechinf: gm_ge_headgear_beret_crew_grn {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_ge_headgear_beret_crew_grn_infantry: gm_ge_headgear_beret_crew_grn {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_ge_headgear_beret_crew_red_militarypolice: gm_ge_headgear_beret_crew_red {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_ge_headgear_beret_crew_red_nbc: gm_ge_headgear_beret_crew_red {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_ge_headgear_beret_crew_red_opcom: gm_ge_headgear_beret_crew_red {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_ge_headgear_beret_crew_bdx_paratrooper: gm_ge_headgear_beret_crew_bdx {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_ge_headgear_beret_crew_bdx_lrrp: gm_ge_headgear_beret_crew_bdx {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_ge_headgear_beret_crew_red_supply: gm_ge_headgear_beret_crew_red {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_ge_headgear_beret_crew_red_signals: gm_ge_headgear_beret_crew_red {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    class gm_ge_bgs_headgear_beret_crew_grn: gm_ge_headgear_beret_crew_grn {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_ge_bgs_headgear_beret_crew_grn_sf: gm_ge_headgear_beret_crew_grn {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    class gm_ge_headgear_hat_beanie_base;
+    class gm_ge_headgear_hat_beanie_crew_blk: gm_ge_headgear_hat_beanie_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    class gm_xx_headgear_headwrap_01_base;
+    class gm_xx_headgear_headwrap_crew_01_oli: gm_xx_headgear_headwrap_01_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_xx_headgear_headwrap_crew_01_m84: gm_xx_headgear_headwrap_01_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_xx_headgear_headwrap_crew_01_grn: gm_xx_headgear_headwrap_01_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_xx_headgear_headwrap_crew_01_smp: gm_xx_headgear_headwrap_01_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_xx_headgear_headwrap_crew_01_blk: gm_xx_headgear_headwrap_01_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_xx_headgear_headwrap_crew_01_flk: gm_xx_headgear_headwrap_01_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class gm_xx_headgear_headwrap_crew_01_trp: gm_xx_headgear_headwrap_01_base {
+        HEARING_PROTECTION_PELTOR;
+    };
+};
